Auxiliary Functions

Contents

Auxiliary Functions

Contents

Index

Auxiliary Functions

Add Finite Rotations

Index

PlateKinematics.Add_FiniteRotations

Auxiliary Functions

Add Finite Rotations

PlateKinematics.Add_FiniteRotations — Function

Add_FiniteRotations(
    FRs1::FiniteRotSph, FRs2::FiniteRotSph, 
    Nsize=1e5::Number, time=nothing::Union{Nothing, Number})

Add_FiniteRotations(
    FRs1::Array{T}, FRs2::Array{T},
    time=nothing::Union{Nothing, Number}) where {T<:FiniteRotSph}

Return the sumation of two Finite Rotations in Spherical coordinates. A specific output :Time field may be passed through the argument time.