Private Functions

Contents

Private Functions

Contents

Index

Private Functions

Covariance-related functions

Mutate Structs

Others

Index

PlateKinematics.CartesianVelocity_toEN

PlateKinematics.ChangeAngle

PlateKinematics.ChangeCovariance

PlateKinematics.ChangeLat

PlateKinematics.ChangeLon

PlateKinematics.ChangeTime

PlateKinematics.CheckCovariance

PlateKinematics.CorrelatedEnsemble3D

PlateKinematics.CovIsZero

PlateKinematics.CovToMatrix

PlateKinematics.GeographicalCoords_toCartesian

PlateKinematics.ReplaceCovariaceEigs

PlateKinematics.ToArray

Private Functions

Covariance-related functions

PlateKinematics.CorrelatedEnsemble3D — Function

CorrelatedEnsemble3D(matrix::Array{N, 2}, Nsize::Int64) where {N<:Float64}

Generates a series of samples [x y z] based on a covariance matrix.

PlateKinematics.CovIsZero — Function

CovIsZero(cov::Covariance)

Check whether all Covariance elements are zero.

PlateKinematics.CheckCovariance — Function

CheckCovariance(covMatrix::Array{N, 2}) where {N<:Float64}

Check whether the covariance matrix yields any negative or imaginary eigenvalue.

PlateKinematics.CovToMatrix — Function

CovToMatrix(FR::Union{FiniteRotSph, FiniteRotCart})

Converts a Finite Rotations Covariance structure [radians^2] to a 3x3 symmetric Matrix [radians^2].

CovToMatrix(EVs::Union{EulerVectorSph, EulerVectorCart})

Convert an Euler Vector Covariance structure [radians²/Myr²] to a 3x3 symmetric Matrix [degrees²/Myr²].

PlateKinematics.ReplaceCovariaceEigs — Function

ReplaceCovariaceEigs(covMatrix::Array{N, 2}) where {N<:Float64}

Checks if a covariance-matrix has negative or imaginary eigenvalues, and replace the diagonal elements (variances) with an average of the positive eigenvalues, while the other elements (covariances) are replaced by zeros.

PlateKinematics.ToArray — Function

ToArray(cov::Covariance)

Convert a Covariance structure to a 1x6 Matrix.

Mutate Structs

PlateKinematics.ChangeLon — Function

ChangeLon(FRs::FiniteRotSph, newLon::Float64)

Change the :Lon (longitude) value of FRs with newLon.

PlateKinematics.ChangeLat — Function

ChangeLat(FRs::FiniteRotSph, newLat::Float64)

Change the :Lat (latitude) value of FRs with newLat.

PlateKinematics.ChangeAngle — Function

ChangeAngle(FRs::FiniteRotSph, newAngle::Float64)

Change the :Angle (angle) value of FRs with newAngle.

PlateKinematics.ChangeTime — Function

ChangeTime(FRs::FiniteRotSph, newTime::Float64)

Change the :Time (rotation age) value of FRs with newTime.

PlateKinematics.ChangeCovariance — Function

ChangeCovariance(
    FRs::FiniteRotSph, newCovariance::Union{Covariance, Array{N}}) where {N<:Float64}

Change the ::Covariance (covariance elements) of FRs with newCovariance.

Others

PlateKinematics.CartesianVelocity_toEN — Function

CartesianVelocity_toEN(pntLon::N, pntLat::N, SVcArray::Array{N, 2}) where {N<:Float64}

Converts a set of velocities from Cartesian to East/North (EN) components.

PlateKinematics.GeographicalCoords_toCartesian — Function

GeographicalCoords_toCartesian(pntLon::Float64, pntLat::Float64, Ht=0.0::Float64)

Converts local geographical coordinates (ellipsoidal) into Cartesian.