Types

This section provides an overview of the types used in PlateKinematics.jl.

Contents

Types

Contents

Index

Types

Covariance

Finite Rotations

Euler Vectors

Surface Velocity

Index

PlateKinematics.Covariance

PlateKinematics.EulerAngles

PlateKinematics.EulerVectorCart

PlateKinematics.EulerVectorSph

PlateKinematics.FiniteRotCart

PlateKinematics.FiniteRotSph

PlateKinematics.Stat

PlateKinematics.SurfaceVelocityVector

Types

Covariance

PlateKinematics.Covariance — Type

struct Covariance

Covariance upper triangle elements. Expressed in radians² for Finite Rotations and in radians²/Myr² for Euler Vectors.

Examples:

julia> PlateKinematics.Covariance()
PlateKinematics.Covariance(0, 0, 0, 0, 0, 0)

julia> PlateKinematics.Covariance(1, 2, 3, 4, 5, 6)
PlateKinematics.Covariance(1, 2, 3, 4, 5, 6)

julia> array = [1, 2, 3, 4, 5, 6];
julia> PlateKinematics.Covariance(array)
PlateKinematics.Covariance(1, 2, 3, 4, 5, 6)

Finite Rotations

PlateKinematics.FiniteRotSph — Type

struct FiniteRotSph

Finite rotation in Spherical coordinates, expressed in degrees.

Fields

Lon::Number: Longitude of the rotation axis in degrees-East

Lat::Number: Latitude of the rotation axis in degrees-North

Angle::Number: Angle of rotation in degrees

Time::Union{Number, Nothing}: Age of rotation in million years

Covariance::Covariance: Covariance in radians²

Examples:

julia> PlateKinematics.FiniteRotSph(30, 20, 10)
PlateKinematics.FiniteRotSph(30, 20, 10, nothing, PlateKinematics.Covariance(0, 0, 0, 0, 0, 0))

julia> array = [30, 20, 10];
julia> PlateKinematics.FiniteRotSph(array)
PlateKinematics.FiniteRotSph(30, 20, 10, nothing, PlateKinematics.Covariance(0, 0, 0, 0, 0, 0))

julia> PlateKinematics.FiniteRotSph(30, 20, 10, 2)
PlateKinematics.FiniteRotSph(30, 20, 10, 2, PlateKinematics.Covariance(0, 0, 0, 0, 0, 0))

julia> array = [1, 2, 3, 4, 5, 6];
julia> PlateKinematics.FiniteRotSph(30, 20, 10, array)
PlateKinematics.FiniteRotSph(30, 20, 10, nothing, PlateKinematics.Covariance(1, 2, 3, 4, 5, 6))

julia> array = [1, 2, 3, 4, 5, 6];
julia> PlateKinematics.FiniteRotSph(30, 20, 10, 2, array)
PlateKinematics.FiniteRotSph(30, 20, 10, 2, PlateKinematics.Covariance(1, 2, 3, 4, 5, 6))

PlateKinematics.FiniteRotCart — Type

struct FiniteRotCart

Finite rotation in Cartesian coordinates, expressed in degrees.

Fields

X::Number: X-coordinate in degrees

Y::Number: Y-coordinate in degrees

Z::Number: Z-coordinate in degrees

TimeRange::Union{Number, Nothing}: Age of rotation in million years

Covariance::Covariance: Covariance in radians²

Examples:

julia> PlateKinematics.FiniteRotCart(1, 2, 3)
PlateKinematics.FiniteRotCart(1, 2, 3, nothing, PlateKinematics.Covariance(0, 0, 0, 0, 0, 0))

julia> array = [30, 20, 10];
julia> PlateKinematics.FiniteRotCart(array)
PlateKinematics.FiniteRotCart(30, 20, 10, nothing, PlateKinematics.Covariance(0, 0, 0, 0, 0, 0))

julia> PlateKinematics.FiniteRotCart(1, 2, 3, 1.5)
PlateKinematics.FiniteRotCart(1, 2, 3, 1.5, PlateKinematics.Covariance(0, 0, 0, 0, 0, 0))

julia> array = [1, 2, 3, 4, 5, 6];
julia> PlateKinematics.FiniteRotCart(30, 20, 10, array)
PlateKinematics.FiniteRotCart(30, 20, 10, nothing, PlateKinematics.Covariance(1, 2, 3, 4, 5, 6))

julia> array = [1, 2, 3, 4, 5, 6];
julia> PlateKinematics.FiniteRotCart(1, 2, 3, 1.5, array)
PlateKinematics.FiniteRotCart(1, 2, 3, 1.5, PlateKinematics.Covariance(1, 2, 3, 4, 5, 6))

PlateKinematics.EulerAngles — Type

struct EulerAngles

Euler angles that describe the rotation around the three main axes on Earth.

Fields

X::Number: Angle of rotation around the X-axis (0N, 0E)

Y::Number: Angle of rotation around the Y-axis (0N, 90E)

Z::Number: Angle of rotation around the Z-axis (90N, 0E)

Examples:

julia> PlateKinematics.EulerAngles(4, 5, 6)
PlateKinematics.EulerAngles(4, 5, 6)

julia> array = [4, 5, 6];
julia> PlateKinematics.EulerAngles(array)
PlateKinematics.EulerAngles(4, 5, 6)

Euler Vectors

PlateKinematics.EulerVectorSph — Type

struct EulerVectorSph

Euler vector in spherical coordinates with the following parameters:

Fields

Lon::Number: Longitude of the Euler pole in degrees-East

Lat::Number: Latitude of the Euler pole in degrees-North

AngVelocity::Number: Angular velocity in degrees/Myr

TimeRange::Union{Matrix, Nothing}: Initial to final age of rotation

Covariance::Covariance: Covariance in radians²/Myr²

Examples:

julia> PlateKinematics.EulerVectorSph(1, 2, 3)
PlateKinematics.EulerVectorSph(1, 2, 3, nothing, PlateKinematics.Covariance(0, 0, 0, 0, 0, 0))

julia> array = [30, 20, 10];
julia> PlateKinematics.EulerVectorSph(array)
PlateKinematics.EulerVectorSph(30, 20, 10, nothing, PlateKinematics.Covariance(0, 0, 0, 0, 0, 0))

julia> array = [1.5 2.5];
julia> length(array) == 2
true
julia> PlateKinematics.EulerVectorSph(30, 20, 10, array)
PlateKinematics.EulerVectorSph(30, 20, 10, [1.5 2.5], PlateKinematics.Covariance(0, 0, 0, 0, 0, 0))

julia> array = [1, 2, 3, 4, 5, 6];
julia> length(array) != 2
true
julia> PlateKinematics.EulerVectorSph(30, 20, 10, array)
PlateKinematics.EulerVectorSph(30, 20, 10, nothing, PlateKinematics.Covariance(1, 2, 3, 4, 5, 6))

julia> array = [1, 2, 3, 4, 5, 6];
julia> PlateKinematics.EulerVectorSph(1, 2, 3, [1.5 2.5], array)
PlateKinematics.EulerVectorSph(1, 2, 3, [1.5 2.5], PlateKinematics.Covariance(1, 2, 3, 4, 5, 6))

PlateKinematics.EulerVectorCart — Type

struct EulerVectorCart

Euler vector in Cartesian coordinates, expressed in degrees/Myr.

Fields

X::Number: X-coordinate in degrees/Myr

Y::Number: Y-coordinate in degrees/Myr

Z::Number: Z-coordinate in degrees/Myr

TimeRange::Union{Matrix, Nothing}: Initial to final age of rotation

Covariance::Covariance: Covariance in radians²/Myr²

Examples:

Same outer Constructor Methods as EulerVectorSph.

Surface Velocity

PlateKinematics.Stat — Type

struct Stat

Mean and standard deviation of a parameter:

Fields

Mean::Number: Mean (average)

StDev::Number: Standard deviation

Examples:

julia> PlateKinematics.Stat(10, 20)
PlateKinematics.Stat(10, 20)

julia> PlateKinematics.Stat([10 20])
PlateKinematics.Stat(10, 20)

PlateKinematics.SurfaceVelocityVector — Type

struct SurfaceVelocityVector

Surface velocity vector components, expressed in mm/yr.

Fields

Lon::Number: Longitude of the surface point in degrees-East

Lat::Number: Latitude of the surface point in degrees-North

EastVel::Union{Number, Stat, Nothing}: East-component of the velocity in mm/yr

NorthVel::Union{Number, Stat, Nothing}: North-component of the velocity in mm/yr

TotalVel::Union{Number, Stat, Nothing}: Total velocity in mm/yr

Azimuth::Union{Number, Stat, Nothing}: Azimuth of the velocity vector as measured clockwise from the North

Examples:

julia> PlateKinematics.SurfaceVelocityVector(10, 20, 4)
PlateKinematics.SurfaceVelocityVector(10, 20, nothing, nothing, 4, nothing)

julia> totalVel = PlateKinematics.Stat(2.5, 2);
julia> PlateKinematics.SurfaceVelocityVector(10, 20, totalVel)
PlateKinematics.SurfaceVelocityVector(10, 20, nothing, nothing, PlateKinematics.Stat(2.5, 2.0), nothing)

julia> PlateKinematics.SurfaceVelocityVector(10, 20, [2.5, 2])
PlateKinematics.SurfaceVelocityVector(10, 20, nothing, nothing, PlateKinematics.Stat(2.5, 2.0), nothing)

julia> PlateKinematics.SurfaceVelocityVector(10, 20, 4, 3)
PlateKinematics.SurfaceVelocityVector(10, 20, 4, 3, nothing, nothing)

julia> PlateKinematics.SurfaceVelocityVector(10, 20, 4, 3, 2)
PlateKinematics.SurfaceVelocityVector(10, 20, 4, 3, 2, nothing)